Откат некоторых правок, которые аффектили юнит-тесты

+++ CodingDojo/games/engine/src/main/java/com/codenjoy/dojo/utils/TestUtils.java
@@ -41,7 +41,7 @@ public class TestUtils {
     }
 
     public static String injectNN(String expected) {
-        int size = (int) Math.sqrt(expected.length()/(double)COUNT_NUMBERS)*COUNT_NUMBERS;
+        int size = (int) Math.sqrt(expected.length()/COUNT_NUMBERS)*COUNT_NUMBERS;
         return inject(expected, size, "\n");
     }
